Create function for user choice and cpu choice
Using the console.log function, I was able to
create the rough algorithm to collect user input
and randomly generate a number which then is assigned
to either 'Rock' 'Paper' or 'Scissors' for the
computer to play.

Will need to create algorithm to analyze and output
results to show winner. Later on, clean up functions
logic and create a web ui to allow a nicer input for user.