Clean up user choice selection logic/algorithm
Made sure that the user's choice functions are
focusing on solving one problem at a time.
Thus, I separated getHumanSelection into two functions,
one for input and localization (make lowecase) and another
to validate and ensure that the input is a valid option.
Furthermore, I was able to clean up the code substantially,
improving efficiency and readability.

Moving on to the computer choice selection function.