Create function to store scores and replay game
After researching while loops, I was able to create
a function to crudely call the user choice prompt and
validation functions. Furthermore, was able to check
and stop loop if user inputs invalid value to ensure
the code doesn't continue.

Furthermore, using the built in confirm argument/function
I was able to end the while loop and confirm if the
user wants to continue.

Lastly, using the global values added, I was able to
increment the scores of whoever wins the round and display
these scores once the user has decided to stop playing.

This concludes this project as the basic framework/logic
was implimented and successfully carried out by the algorithms
I created. Anything else added to this will be for more
advanced ideas or testing.

Going back to TOP to continue the foundations course.